refactor(image): image 패키지의 application안에 있던 domain 패키지 바깥으로 뺴냄
- 변경 전 image/application/domain
- 변경 후 image/domain & image/application
