+++ backend/src/main/java/com/wooteco/nolto/image/application/ImageService.java
@@ -6,8 +6,8 @@ import com.wooteco.nolto.exception.BadRequestException;
 import com.wooteco.nolto.exception.ErrorType;
 import com.wooteco.nolto.exception.InternalServerErrorException;
 import com.wooteco.nolto.image.application.adapter.ImageHandlerAdapter;
-import com.wooteco.nolto.image.application.domain.ProcessedImage;
-import com.wooteco.nolto.image.application.domain.repository.ImageRepository;
+import com.wooteco.nolto.image.domain.ProcessedImage;
+import com.wooteco.nolto.image.domain.repository.ImageRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
