+++ backend/src/main/java/com/wooteco/nolto/image/application/adapter/ImageHandlerAdapter.java
@@ -1,6 +1,6 @@
 package com.wooteco.nolto.image.application.adapter;
 
-import com.wooteco.nolto.image.application.domain.ProcessedImage;
+import com.wooteco.nolto.image.domain.ProcessedImage;
 import org.apache.commons.io.FilenameUtils;
 
 import java.io.File;
