--- backend/src/main/java/com/wooteco/nolto/image/application/domain/repository/ImageRepository.java
@@ -1,8 +1,0 @@
-package com.wooteco.nolto.image.application.domain.repository;
-
-import com.wooteco.nolto.image.application.domain.ProcessedImage;
-
-public interface ImageRepository {
-
-    String save(ProcessedImage processedImage);
-}
