+++ backend/src/main/java/com/wooteco/nolto/image/domain/repository/ImageRepository.java
@@ -1,0 +1,8 @@
+package com.wooteco.nolto.image.domain.repository;
+
+import com.wooteco.nolto.image.domain.ProcessedImage;
+
+public interface ImageRepository {
+
+    String save(ProcessedImage processedImage);
+}
